Portfolio
 ### Hello World

This is my home page and! My Name is Nathaniel Cruz and I am a student at [Cal State Fullerton](http://www.fullerton.edu/) and my major is Electrical Engineering.

## Computer Science Projects

My GitHub page is http://github.com/NathanC5992.

### CPSC 120

Lab 3:

Lab 03 part 1 - This Lab would one of my favorites due to the conversion of pizza size to the overall price. It really interested me when I was able to put in the size that the price would be responded on the terminal. I felt like I was coding a calculator which was very fun for me.

Lab 5:

Lab 05 part 2 - Being someone who enjoys going to the casino to play black jack this lab was very interesting to me. I enjoyed how I was able to make the computer play the game of black jack with me and keep score. It was fun because I actually wasn’t losing any money in the process.

Lab 9:

Lab 09 part 1 - This lab was very interesting to myself because it was basically making a dictionary. In order to find the meaning of the words we had to implement the functions to search for each meaning. Although it was difficult I found it very satisfying to complete.